Add queryHasMany method to the store
