+++ app/routes/preprints/provider/moderation.js
@@ -1,10 +1,5 @@
-import { bind } from '@ember/runloop';
 import { inject as service } from '@ember/service';
 import Route from '@ember/routing/route';
-import PromiseProxyMixin from '@ember/object/promise-proxy-mixin';
-import ArrayProxy from '@ember/array/proxy';
-import $ from 'jquery';
-import { Promise as EmberPromise } from 'rsvp';
 /**
  * @module ember-osf-reviews
  * @submodule routes
@@ -25,7 +20,7 @@ export default Route.extend({
 
     model(params) {
         const provider = this.modelFor('preprints.provider');
-        return this.query(provider, 'preprints', {
+        return this.get('store').queryHasMany(provider, 'preprints', {
             'filter[reviews_state]': params.status,
             'meta[reviews_state_counts]': true,
             sort: params.sort,
@@ -56,38 +51,4 @@ export default Route.extend({
             statusCounts: response.meta.reviews_state_counts,
         };
     },
-
-    // TODO: add this to osf-model, remove ember-data-has-many-query dependency?
-    query(model, propertyName, params) {
-        const reference = model.hasMany(propertyName);
-        const promise = new EmberPromise((resolve, reject) => {
-            // HACK: ember-data discards/ignores the link if an object on the belongsTo side
-            // came first. In that case, grab the link where we expect it from OSF's API
-            const url = reference.link() || model.get(`links.relationships.${propertyName}.links.related.href`);
-            if (url) {
-                $.ajax(url, {
-                    data: params,
-                    xhrFields: {
-                        withCredentials: true,
-                    },
-                }).catch(reject)
-                    .then(bind(this, this._queryDone, resolve));
-            } else {
-                reject(`Could not find a link for '${propertyName}' relationship`);
-            }
-        });
-
-        const ArrayPromiseProxy = ArrayProxy.extend(PromiseProxyMixin);
-        return ArrayPromiseProxy.create({ promise });
-    },
-
-    _queryDone(resolve, payload) {
-        const store = this.get('store');
-        store.pushPayload(payload);
-        const records = payload.data.map(datum => store.peekRecord(datum.type, datum.id));
-        records.meta = payload.meta;
-        records.links = payload.links;
-        resolve(records);
-    },
-
 });
